requirements.txt
asgiref==3.6.0
Django==4.2
django-braces==1.15.0
Pillow==9.5.0
sqlparse==0.4.4
